feat: Update aries-framework-go, only use new storage interface now
