--- .github/workflows/newstorage-couchdb.yml
@@ -1,51 +1,0 @@
-#
-# Copyright SecureKey Technologies Inc. All Rights Reserved.
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-name: storage-couchdb
-on:
-  push:
-    paths:
-      - 'component/newstorage/couchdb/**'
-  pull_request:
-    paths:
-      - 'component/newstorage/couchdb/**'
-jobs:
-  linter:
-    name: Go linter
-    timeout-minutes: 10
-    env:
-      LINT_PATH: component/newstorage/couchdb
-    runs-on: ubuntu-18.04
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Checks linter
-        timeout-minutes: 10
-        run: make lint
-  unitTest:
-    name: Unit test
-    runs-on: ubuntu-18.04
-    timeout-minutes: 15
-    env:
-      UNIT_TESTS_PATH: component/newstorage/couchdb
-    steps:
-      - name: Setup Go 1.15
-        uses: actions/setup-go@v2
-        with:
-          go-version: 1.15
-        id: go
-
-      - uses: actions/checkout@v2
-
-      - name: Run unit test
-        timeout-minutes: 15
-        run: make unit-test
-
-      - name: Upload coverage to Codecov
-        timeout-minutes: 10
-        if: github.repository == 'hyperledger/aries-framework-go-ext'
-        uses: codecov/codecov-action@v1.0.14
-        with:
-          file: ./coverage.out
