+++ component/didcomm/transport/amqp/go.mod
@@ -7,10 +7,10 @@ go 1.15
 
 require (
 	github.com/cenkalti/backoff v2.2.1+incompatible
-	github.com/google/uuid v1.1.1
-	github.com/hyperledger/aries-framework-go v0.1.5-0.20201017112511-5734c20820a9
+	github.com/google/uuid v1.1.2
+	github.com/hyperledger/aries-framework-go v0.1.6-0.20210225161605-5a3ea609e830
 	github.com/ory/dockertest/v3 v3.6.2
 	github.com/pkg/errors v0.9.1
 	github.com/streadway/amqp v1.0.0
-	github.com/stretchr/testify v1.6.1
+	github.com/stretchr/testify v1.7.0
 )
