--- component/newstorage/couchdb/testdata/single-node.ini
@@ -1,20 +1,0 @@
-#
-# Copyright SecureKey Technologies Inc. All Rights Reserved.
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-
-# Explicitly set single_node=true when not in cluster mode.
-# This ensures our single CouchDB node creates all system databases for our tests, including _users.
-# References:
-#  - https://docs.couchdb.org/en/3.1.0/config/couchdb.html#couchdb/single_node
-#  - https://github.com/apache/couchdb-docker#no-system-databases-until-the-installation-is-finalized
-#  - https://github.com/apache/couchdb-docker/issues/54#issuecomment-643818998
-[couchdb]
-single_node=true
-
-[admins]
-admin = -pbkdf2-02c79f17107e107b4954ee8754a66f8bb5751fcb,c95bbbe947dddb0b70142d4c1d0bd495,10
-
-[couch_httpd_auth]
-secret = ed2e8603452743d0b70d042c253f25e3
