+++ component/storage/couchdb/go.mod
@@ -1,24 +1,23 @@
 // Copyright SecureKey Technologies Inc. All Rights Reserved.
 //
 // SPDX-License-Identifier: Apache-2.0
-module github.com/hyperledger/aries-framework-go-ext/component/storage/couchdb
+module github.com/hyperledger/aries-framework-go-ext/component/newstorage/couchdb
 
 go 1.15
 
 require (
 	github.com/cenkalti/backoff v2.2.1+incompatible
-	github.com/flimzy/diff v0.1.7 // indirect
-	github.com/flimzy/testy v0.1.17 // indirect
-	github.com/go-kivik/couchdb v2.0.0+incompatible
-	github.com/go-kivik/kivik v2.0.0+incompatible
-	github.com/go-kivik/kiviktest v2.0.0+incompatible // indirect
-	github.com/google/uuid v1.1.2
-	github.com/gopherjs/gopherjs v0.0.0-20200217142428-fce0ec30dd00 // indirect
-	github.com/hyperledger/aries-framework-go v0.1.5-0.20201028195746-556cee009e20
-	github.com/hyperledger/aries-framework-go-ext/test/component/storage v0.0.0
-	github.com/ory/dockertest/v3 v3.6.2
-	github.com/stretchr/testify v1.6.1
-	gitlab.com/flimzy/testy v0.3.2 // indirect
+	github.com/go-kivik/couchdb/v3 v3.2.6
+	github.com/go-kivik/kivik/v3 v3.2.3
+	github.com/google/go-cmp v0.5.2 // indirect
+	github.com/google/uuid v1.2.0
+	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210225161605-5a3ea609e830
+	github.com/hyperledger/aries-framework-go/test/component v0.0.0-20210225161605-5a3ea609e830
+	github.com/kr/pretty v0.1.0 // indirect
+	github.com/ory/dockertest/v3 v3.6.3
+	github.com/stretchr/testify v1.7.0
+	golang.org/x/net v0.0.0-20210119194325-5f4716e94777 // indirect
+	golang.org/x/sys v0.0.0-20201211090839-8ad439b19e0f // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
-
-replace github.com/hyperledger/aries-framework-go-ext/test/component/storage => ../../../test/component/storage
