+++ component/storage/mongodb/go.mod
@@ -7,7 +7,7 @@ go 1.15
 
 require (
 	github.com/cenkalti/backoff v2.2.1+incompatible
-	github.com/hyperledger/aries-framework-go v0.1.4
+	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210225161605-5a3ea609e830
 	github.com/ory/dockertest/v3 v3.6.2
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.6.1
