+++ component/storage/mysql/go.mod
@@ -1,17 +1,21 @@
 // Copyright SecureKey Technologies Inc. All Rights Reserved.
 //
 // SPDX-License-Identifier: Apache-2.0
-module github.com/hyperledger/aries-framework-go-ext/component/storage/mysql
+module github.com/hyperledger/aries-framework-go-ext/component/newstorage/mysql
 
 go 1.15
 
 require (
 	github.com/cenkalti/backoff/v4 v4.1.0
 	github.com/go-sql-driver/mysql v1.5.0
-	github.com/hyperledger/aries-framework-go v0.1.5-0.20201028195746-556cee009e20
-	github.com/hyperledger/aries-framework-go-ext/test/component/storage v0.0.0
+	github.com/google/go-cmp v0.5.2 // indirect
+	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210208223040-55e2734dc08e
+	github.com/hyperledger/aries-framework-go/test/component v0.0.0-20210208223040-55e2734dc08e
+	github.com/kr/pretty v0.1.0 // indirect
 	github.com/ory/dockertest/v3 v3.6.2
 	github.com/stretchr/testify v1.6.1
+	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
+	golang.org/x/sys v0.0.0-20201009025420-dfb3f7c4e634 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
-
-replace github.com/hyperledger/aries-framework-go-ext/test/component/storage => ../../../test/component/storage
