+++ component/storage/mysql/store_test.go
@@ -8,24 +8,26 @@ package mysql_test
 import (
 	"database/sql"
 	"fmt"
+	"log"
 	"os"
 	"testing"
 	"time"
 
 	"github.com/cenkalti/backoff/v4"
 	"github.com/go-sql-driver/mysql"
+	"github.com/hyperledger/aries-framework-go/spi/storage"
+	commontest "github.com/hyperledger/aries-framework-go/test/component/storage"
 	dctest "github.com/ory/dockertest/v3"
 	dc "github.com/ory/dockertest/v3/docker"
 	"github.com/stretchr/testify/require"
 
-	. "github.com/hyperledger/aries-framework-go-ext/component/storage/mysql"
-	common "github.com/hyperledger/aries-framework-go-ext/test/component/storage"
+	. "github.com/hyperledger/aries-framework-go-ext/component/newstorage/mysql"
 )
 
 type mysqlLogger struct{}
 
 // Print ignores MySQL logs.
-func (*mysqlLogger) Print(_ ...interface{}) {}
+func (*mysqlLogger) Print(...interface{}) {}
 
 const (
 	dockerMySQLImage = "mysql"
@@ -50,6 +52,9 @@ func TestMain(m *testing.M) {
 		},
 	})
 	if err != nil {
+		log.Println(`Failed to start MySQL Docker image.` +
+			` This can happen if there is a MySQL container still running from a previous unit test run.` +
+			` Try "docker ps" from the command line and kill the old container if it's still running.`)
 		panic(fmt.Sprintf("run with options: %v", err))
 	}
 
@@ -83,8 +88,8 @@ func checkMySQL() error {
 	}, backoff.WithMaxRetries(backoff.NewConstantBackOff(time.Second), retries))
 }
 
-func TestSqlDBStore(t *testing.T) {
-	t.Run("Test sql multi store put and get", func(t *testing.T) {
+func TestSQLDBStore(t *testing.T) {
+	t.Run("Test SQL open store", func(t *testing.T) {
 		prov, err := NewProvider(sqlStoreDBURL, WithDBPrefix("prefixdb"))
 		require.NoError(t, err)
 
@@ -92,13 +97,11 @@ func TestSqlDBStore(t *testing.T) {
 		require.Error(t, err)
 		require.Equal(t, err.Error(), "store name is required")
 	})
-
 	t.Run("Test wrong url", func(t *testing.T) {
 		_, err := NewProvider("root:@tcp(127.0.0.1:45454)/")
 		require.Error(t, err)
 		require.Contains(t, err.Error(), "failure while pinging MySQL")
 	})
-
 	t.Run("Test sql db store failures", func(t *testing.T) {
 		prov, err := NewProvider("")
 		require.Error(t, err)
@@ -114,7 +117,6 @@ func TestSqlDBStore(t *testing.T) {
 		require.Error(t, err)
 		require.Contains(t, err.Error(), "failure while pinging MySQL")
 	})
-
 	t.Run("Test sqlDB multi store close by name", func(t *testing.T) {
 		prov, err := NewProvider(sqlStoreDBURL, WithDBPrefix("prefixdb"))
 		require.NoError(t, err)
@@ -148,17 +150,9 @@ func TestSqlDBStore(t *testing.T) {
 			require.NoError(t, e)
 			require.NotNil(t, store)
 
-			e = prov.CloseStore(name)
-			require.NoError(t, e)
-
-			// try to get after provider is closed
-			_, err = store.Get(commonKey)
-			require.Error(t, err)
+			err = store.Close()
+			require.NoError(t, err)
 		}
-
-		// try to close non existing db
-		err = prov.CloseStore("store_x")
-		require.EqualError(t, err, ErrStoreNotFound.Error())
 
 		err = prov.Close()
 		require.NoError(t, err)
@@ -167,11 +161,55 @@ func TestSqlDBStore(t *testing.T) {
 		err = prov.Close()
 		require.NoError(t, err)
 	})
+	t.Run("Flush", func(t *testing.T) {
+		prov, err := NewProvider(sqlStoreDBURL)
+		require.NoError(t, err)
+
+		store, err := prov.OpenStore("storename")
+		require.NoError(t, err)
+
+		err = store.Flush()
+		require.NoError(t, err)
+	})
+}
+
+func TestNotImplementedMethods(t *testing.T) {
+	t.Run("Not implemented methods", func(t *testing.T) {
+		prov, err := NewProvider(sqlStoreDBURL)
+		require.NoError(t, err)
+
+		err = prov.SetStoreConfig("", storage.StoreConfiguration{})
+		require.EqualError(t, err, "not implemented")
+
+		_, err = prov.GetStoreConfig("")
+		require.EqualError(t, err, "not implemented")
+
+		openStores := prov.GetOpenStores()
+		require.Nil(t, openStores)
+
+		store, err := prov.OpenStore("storename")
+		require.NoError(t, err)
+
+		_, err = store.GetTags("")
+		require.EqualError(t, err, "not implemented")
+
+		_, err = store.GetBulk()
+		require.EqualError(t, err, "not implemented")
+
+		_, err = store.Query("")
+		require.EqualError(t, err, "not implemented")
+
+		err = store.Batch(nil)
+		require.EqualError(t, err, "not implemented")
+	})
 }
 
 func TestSqlDBStore_Common(t *testing.T) {
-	prov, err := NewProvider(sqlStoreDBURL)
+	provider, err := NewProvider(sqlStoreDBURL)
 	require.NoError(t, err)
 
-	common.TestAll(t, prov)
+	commontest.TestPutGet(t, provider)
+	commontest.TestStoreDelete(t, provider)
+	commontest.TestStoreClose(t, provider)
+	commontest.TestProviderClose(t, provider)
 }
