+++ component/vdr/sidetree/go.mod
@@ -6,8 +6,8 @@ module github.com/hyperledger/aries-framework-go-ext/component/vdr/sidetree
 go 1.15
 
 require (
-	github.com/hyperledger/aries-framework-go v0.1.6-0.20210120000618-bdf82385e9df
+	github.com/hyperledger/aries-framework-go v0.1.6-0.20210225161605-5a3ea609e830
 	github.com/square/go-jose/v3 v3.0.0-20200630053402-0a67ce9b0693
-	github.com/stretchr/testify v1.6.1
+	github.com/stretchr/testify v1.7.0
 	github.com/trustbloc/sidetree-core-go v0.1.6-0.20201217192009-0d2b4436912f
 )
