+++ component/vdr/trustbloc/go.mod
@@ -7,9 +7,9 @@ go 1.15
 
 require (
 	github.com/bluele/gcache v0.0.0-20190518031135-bc40bd653833
-	github.com/hyperledger/aries-framework-go v0.1.6-0.20210120000618-bdf82385e9df
+	github.com/hyperledger/aries-framework-go v0.1.6-0.20210225161605-5a3ea609e830
 	github.com/hyperledger/aries-framework-go-ext/component/vdr/sidetree v0.0.0-20210121210840-ee9984a4579c
 	github.com/sirupsen/logrus v1.4.2
 	github.com/square/go-jose/v3 v3.0.0-20200630053402-0a67ce9b0693
-	github.com/stretchr/testify v1.6.1
+	github.com/stretchr/testify v1.7.0
 )
