--- test/component/storage/go.mod
@@ -1,12 +1,0 @@
-// Copyright SecureKey Technologies Inc. All Rights Reserved.
-//
-// SPDX-License-Identifier: Apache-2.0
-module github.com/hyperledger/aries-framework-go-ext/test/component/storage
-
-go 1.15
-
-require (
-	github.com/google/uuid v1.1.2
-	github.com/hyperledger/aries-framework-go v0.1.5-0.20201028195746-556cee009e20
-	github.com/stretchr/testify v1.6.1
-)
